+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @webscopeio/react-firebase-messenger
 
+## 0.0.5
+
+### Patch Changes
+
+- Update: replace ramda with rambda (+ts fixes)
+
 ## 0.0.4
 
 ### Patch Changes